+++ src/org/ods/orchestration/usecase/LeVADocumentUseCase.groovy
@@ -1597,7 +1597,7 @@ class LeVADocumentUseCase extends DocGenUseCase {
         if (this.project.historyForDocumentExists(documentName)) {
             return this.project.getHistoryForDocument(documentName)
         } else {
-            def documentType = documentName.split('-').first()
+            def documentType = LeVADocumentUtil.getTypeFromName(documentName)
             def jiraData = this.project.data.jira as Map
             def environment = this.computeSavedDocumentEnvironment(documentType)
             def docHistory = new DocumentHistory(this.steps, ServiceRegistry.instance.get(Logger), environment, documentName)
@@ -1617,7 +1617,7 @@ class LeVADocumentUseCase extends DocGenUseCase {
     protected String computeSavedDocumentEnvironment(String documentType) {
         def environment = this.project.buildParams.targetEnvironmentToken
         if (this.project.isWorkInProgress) {
-            environment = ['D', 'Q', 'P'].find { env ->
+            environment = Environment.values().collect { it.toString() }.find { env ->
                 LeVADocumentScheduler.ENVIRONMENT_TYPE[env].containsKey(documentType)
             }
         }
@@ -1742,7 +1742,8 @@ class LeVADocumentUseCase extends DocGenUseCase {
                 // The document, or a new version of it, has already been created in this same pipeline run.
                 version = this.project.getDocumentVersionFromHistories(doc)
                 if (!version) {
-                    def trackingIssues =  this.getDocumentTrackingIssuesForHistory(doc, ['D', 'Q', 'P'])
+                    def envs = Environment.values().collect { it.toString() }
+                    def trackingIssues =  this.getDocumentTrackingIssuesForHistory(doc, envs)
                     version = this.jiraUseCase.getLatestDocVersionId(trackingIssues)
                     if (this.project.isWorkInProgress || docIsCreatedInTheEnvironment(doc)) {
                         // Either this is a developer preview or