+++ test/groovy/org/ods/orchestration/usecase/LeVADocumentUseCaseSpec.groovy
@@ -5,6 +5,7 @@ import groovy.util.logging.Log
 import org.apache.commons.io.FileUtils
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
+import org.ods.services.ServiceRegistry
 import spock.lang.Unroll
 
 import org.ods.services.JenkinsService
@@ -62,6 +63,7 @@ class LeVADocumentUseCaseSpec extends SpecHelper {
         pdf = Mock(PDFUtil)
         sq = Mock(SonarQubeUseCase)
         logger = Mock(Logger)
+        ServiceRegistry.instance.add(Logger, logger)
         bbt = Mock(BitbucketTraceabilityUseCase)
         usecase = Spy(new LeVADocumentUseCase(project, steps, util, docGen, jenkins, jiraUseCase, junit, levaFiles, nexus, os, pdf, sq, bbt))
         project.getOpenShiftApiUrl() >> 'https://api.dev-openshift.com'